mon: fix warning

+++ src/mon/Monitor.cc
@@ -640,7 +640,7 @@ void Monitor::handle_subscribe(MMonSubscribe *m)
   // ???
 
   if (reply)
-    messenger->send_message(new MMonSubscribeAck(g_conf.mon_subscribe_interval),
+    messenger->send_message(new MMonSubscribeAck((int)g_conf.mon_subscribe_interval),
 			    m->get_source_inst());
 
   s->put();
